move item.ts duck into proper folder
