+++ src/redux/ducks/items/index.ts
@@ -1,4 +1,4 @@
-import { IAction } from '../types';
+import { IAction } from '../../types';
 // interfaces
 export interface IItem {
   tags: string[];
