+++ src/redux/sagas/rootSaga.ts
@@ -1,6 +1,6 @@
 import { takeLatest, all } from 'redux-saga/effects';
 import { handleGetItems } from './item';
-import { GET_ITEMS } from '../ducks/item';
+import { GET_ITEMS } from '../ducks/items';
 import { GET_COMPANIES } from '../ducks/company';
 import { handleGetCompanies } from './company';
